+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Pre-release (2019-03-31)
+
+* Translators for schema and codelists.
+
 ## Pre-release (2019-02-22)
 
 * Initial setup with extractors for schema and codelists.